First cut at node naming.
The node names are of the form 'sirius-$HOST_NAME:$AKKA_PORT', and the
node name is currently exposed only through a JMX hook.  In order to
register and unregister the mbean, I modified Sirius to have an explicit
supervisor actor.  Registering and unregistering the mbean is done as
part of that actor's lifecyle.

Change-Id: I523a640dfb6f6460a637d006fcb8006dc8b7cc5e

A few merge conflicts I missed.

Need to actually pass the LogWriter through to the
SiriusPersistanceActor.

Change-Id: Ic29023078c6417262672daf5df773826acb9e66d