Sirius standalone build support
Copied mostly from octave, where it was quite useful.

A build script to assemble a tarball with sirius and dependencies,
as well as a launcher script that will appropriately configure the
classpath and launch a scala shell to interact with sirius in.

This will be useful for testing and demoing.

Change-Id: Id0b0afd2c1b59c24b80b2665f0282a6f42beaf8d