Extract creation of SiriusSupervisor
Not sure if this is as clean as it can get, but it's a start.  By
pushing the instantiation stuff into a method we can override it and
return our own probe.  This may be the natural progression to using
a cake like pattern, it may not.

Also, now we can probably start eliminating some of the constructor
arguments.

Change-Id: I5d27ba1749f628bf379192e6f56e26496d33e3b1