Make value private
I think this means that we can only pattern match a SiriusResult
from within a SiriusResult.  However, that is all that we are doing
with it for the time being, so it should be ok.  We can always
become less restrictive.  This change is more to encourage safe
usage from the Java API and avoid implementation leak.

Change-Id: Ibfaa56996ad773d48195613a1683f46364b60f41