Add SiriusState Agent for shared Agent state.
SiriusSupervisor now starts out in an uninitialized state, and ignores messages while in that state.
When each message is received, it checks if the child Actors (only the persistence actor, for now) are
in the Initialized state in the SiriusState Agent.  If they are, the SiriusSupervisor 'becomes'
Initialized, and begins to respond to messages.

Several iTests needed to be updated to wait for the SiriusSupervisor to transition to the Initialized
state before proceeding.

Change-Id: Iba6a1856dd6a86cad2837930559a0b633991e8bc