Add GetRandomMember functionality in MembershipActor
For new node bootstrapping. We need the ability to fairly choose a member
of the cluster so that we can ask for their logs.  This is a naive
implementation, simply getting a random member, not checking for
liveness or anything.

patchset2: a GetRandomMember message to a MembershipActor will now NOT
return a reference to that same actor.  That is, if we ask a
MembershipActor for a random actor, it will not return itself.  If no
viable actor exists to return, None is returned.

patchset6: SiriusInfo is now wired through the SiriusSupervisor and
passed to the MembershipActor.  GetRandomMember is now guaranteed to
finish with at most one random get from the map.

Change-Id: I4f10f2d339765cd3c26b9740a150109413881b38