Add interface for reading lines of SiriusLog
Added LogLinesSource, trait extended by SiriusLog, that defines a method
for retrieving lines of the file.  This should add a getLines(range) in
the future.  With this, the interface can be used to access either
single log file or multiple, if the log is split up.  Used right now for
the LogSending/Receiving chunk of bootstrapping a new node.

fixed: added itest for SiriusLog's getLines

Change-Id: I874a933a43c70caab001b9a4a00aae4405200391