Create MemberInfo, encaps Option[MembershipData]
For sending refs to members around as messages.  The response to
GetRandomMember really needs to be an Option[MembershipData].
However, receiving and handling an Option in another actor is
painful, particularly for the None case, since you don't know
what it's None of.  Wrapping that up as MemberInfo.

Fixed: moved MemberInfo out of package.scala.  Would be nice if it could
extend MembershipMessage, but it can't extend MembershipMessage and be
matched by classes in another package as long as MM is in a package
object.

Change-Id: Id1426bf5181d8c57ba2d47eaea8b478dd1f3217e