LogReceiver writes received data to Pers Actor
LogReceiver deserializes data received from Sender and sends it over to
the PersistenceActor for storage in memory and on disk.  Added a ref to
the persistenceActor and a serializer/deserializer to the
LogRequestActor to enable this, and propagated that change up where
necessary.

Fixed: Added scaladoc comments to LogSending/Receiving Actors.  Removed
offending type specifications in case statements that were causing the
erasure warnings.  Removed TODO in LogSendingActor by rewriting
gatherLines all functional-like.

Fixed2: Added comments to wrapActorWithMockedSupervisor.
Fixed3: Moved wrapActorWithMockedSupervisor to TestHelper object.

Change-Id: I69bee98bac5c3caa0df9cda4b346ef4a47f7d848