Config change to enable remote messaging
Akka seems to have trouble with remote messaging when the netty path is
set up using a hostname, instead of an ip address or the any flag (empty
hostname).  Setting this to empty to get remote messages to be properly
fielded until we get actual cluster/messaging config ironed out.

Change-Id: I3b2e93533cc57f54582be5a83ddc303f684bb18e