Replica tests
I'm not 100% sure of what I think about this.  On one hand it's nice
to mock stuff because the logic gets pretty complex, on the other
hand the tests can be a little misleading.  It may make sense to
just call through.  More boiler plate, but what's happening is
more obvious.

Change-Id: I29db5691772be142d6017d95f59f01ee86ae83cd