Use OrderedEvent instead of raw lines or LogData
This is the result of squashing a collection of commits which
cumulatively draw the line of abstraction between the write ahead
log layer and the protocol layer.  Previously that line was
quite blurry.  A fair deal of code was also able to be eliminated
by this commit.

WAL interface only uses OrderedEvents now

Change-Id: I731a5fe96163df9d390e5d579e99dc644adda310

Drop LogData class and rename LogLinesSource

Change-Id: I42595a9d42087620f90512eb81ef211714e9dc78

Rename variables and classes to reflect refactoring

Change-Id: Ic9c1e89804ca1b36d1be6dbeb7b44051aa96eb97