Remove re-proposing for now.
Remove the bit of paxos that will re-propose proposals which lost the
election for their slot.  This means that those will just time out on
the client, which means that we can't really be writing many concurrent
requests at the moment because most of them will fail.  We should add
this back in later.

Change-Id: Iea906966129834ee6484110aeb652d394b11f710