Make starting slot for paxoses configurable
Had the epiphany that the current implementations works when starting
from an empty log or an established cluster, but not when all nodes
are started from scratch with an existing log, such that their seq
number may be later.

We will have ton consider what number we tell the paxos system to
start with, if we blindly go with the number from the end of the
write ahead log replay we may wind up rejecting requests which are
within the current allowed half hour window.  That however isn't
necessarily a bad thing.

Change-Id: Ibbf58c653d08c57461bf4d85a5e7b74b31bcc4a4