Membership as a Set[ActorRef]
As discussed during scrum:
 - while membership is a Map[String, MembershipData] all we ever
   really do with membership/are planning to do with it in the
   forseeable future is use the valueset
 - representing all of the top level supervisor's children in the
   membership presents challenges with initialization

This commit converts all of the membership data to be stored in a
Set[ActorRef], which is exactly how we plan on using it. A sort of
discussed change, but one that wasn't agreed on, is using akka
paths/addresses for identifying members.  While developing this
seemed to be the natural progression- the existing name shortening
was from before we did any akka work, and was a convention that
wasn't based on much.  I think we use the full actor paths for now,
then derive a shortening scheme later to make human interaction
shorter/more pleasant, though that's just a matter of opinion.

Also, another change was to have the MembershipHelper#getRandomMember
method take an ActorRef instead of a String, this also seemed most
natural.  To make this work the LogRequestActor has a reference to
the top level local sirius node actor.

Change-Id: I7760126bbbcf54225dee53a7374fa795ceae8232