Wire up paxos, in the most dead simple of ways
Introduce SiriusPaxosAdapter to wrap the intricacies of properly
applying decisions and starting up the paxos subsystem.

The most difficult thing here is that our current log format expects
events to arrive explicitly in sequence number order.  Because of
this we may not commit an event to the log until it is actually its
time.  See the comments in SiriusPaxosAdapter for more information.

Note that when using paxos we need to submit requests using
PaxosSup.Submit.  While it would be nice to maintain the same api
as we used before, that was just a placeholder.  The PaxosSup.Submit
message is so that we can add extra parameters when submitting a
request to paxos (such as retries), and maintain some sort of type
safety.

Change-Id: I2b1e42de85b7ded05af998424a1990d3877229a4