More intuitive shutdown
Remove the boolean for shutting down the ActorSystem.  The passed in
ActorSystem is not destroyed (which is the obvious behavior, because
our objects should not be in the business of modifying things which
don't belong to them).  However, if we create Sirius using the
factory method which also spins up an ActorSystem, the ActorSystem
will be shutdown on startup.  It may make sense to move the factory
style stuff out of SiriusImpl into its own class.

Change-Id: Ifb20ee9b64b1b0a919dbb7c7a3603761298403c3