Buffer out-of-order writes in persistence actor
Simple start for buffering out-of-order messages in the persistence
layer and writing them to disk when possible.  Doesn't do any of
the intelligence we'll need for making this work in the real world
(asking neighbors to fill in gaps after a while, etc).  Will write
in-order messages, buffer not-yet-ready messages, and ignore messages
for sequence numbers already written.

Change-Id: I2ee0139c1df1fb03cc0ef73ae8aa22eb969eea56