More performant checksumming
Result of some digging MBL and I did in the bowels of Scala.

While foldLeft was elegant, it resulted in a lot of autoboxing due
to the implementation of foldLeft.  This version, while not as sexy
is more performant (tested to be about 60% faster) and doesn't
generate all of the extra Long objects. Alas, but Java, GC, who
cares?  10GB * at least 2 autoboxed longs per character = 20 billion
extra objects created when scanning the file.  GC does not like that.

Change-Id: Icecd40ba3b5fae8054e9cae2cb7ba6c20024d6a0