Improve log parsing line splitting performance
Create function quickSplit which splits a string on a single char and
returns a list of all of the tokens in reverse order.  While we could
reverse the list of tokens on completion, this is all internal, and
doesn't seem to be too confusing for the extra objects that the
alternative would create.

Performance increase is roughly 95%

Change-Id: I5b89fdbf075079cdc5808eb0048c0e8c9a9a43a8