Add a SiriusConfiguration object.
There's a bunch of configuration cruft that's been polluting
our factory methods.  Factor it out into a seperate config
object and use that instead of a bunch of factory method args.

The old factory methods are still there but marked as depricated,
once we updated the XinfityTV api to use the new factory method
we can ditch them.

Also updated the tests to not use depricated factory methods.

Change-Id: Iaf21d6519e4a673fb086ab0c864903c6b9a67565