Extract Paxos Decision ordering logic into extra Actor
To do this we moved the core logic of SiriusPaxosAdapter to
PaxosStateBridge, and have SiriusPaxosAdapter forward all decisions
as messages to this actor.

PaxosStateBridge is responsible for ordering Decisions, responding to
them, and making sure they are only sent to the persistence layer
in order.  It really does what it says, bridges the madness of Paxos
and the calmness of the state.

SiriusPaxosAdapter is now more scaffolding.  We use a cool trick,
though not really a trick, where we pass in an implicit ActorContext,
this essentially says, "hey call this from within an actor", and we
use this to build up the necessary Paxos goodies, while not breaking
the naiive adapter ordering jawn.

Change-Id: Ibd9d0b1e95bd816b1cc5488c549c341bb3c8c608