Use a random ID in Leader Ballots
This is because of the situation we discovered yesterday where all
the Leaders, despite being on different machines in different
systems, all had the same ActorPath, because they saw themselves as
LocalActorRefs.

This is good for now, but it kind of sucks because it's not so easy
to trace an ID back to a node, though we haven't realy had to do that
yet anyway... If we do get to that point, there's some code here
(http://doc.akka.io/docs/akka/snapshot/scala/serialization.html) that
will allow us to do that.

Change-Id: I95c8cc527472c6ee07132cbf2a066207a2596f91