Request gaps in eventBuffer from another node
First step to catchup.  When receiving a RequestGaps message,
PaxosStateBridge will find all current gaps in its eventBuffer and
request that those gaps be filled by some other (random) node.

This will need to be run on a scheduler.

As a sidenote: there is currently a large disparity between how long it
will take for a node to respond when the requested sequence is found in
the writeCache and when it needs to be retrieved from disk.  If that
stands, logic to prevent the same gap from being requested many times
will need to be put in place, either here, in the logRequestActor, or
somewhere else -- which gets pretty hairy, so we should probably find
a way to speed up lookup from the log directly.

Change-Id: I1cd84ad0d4210170de15049a5a2f0d3200a54757