Parameterize LogRequest messages with targetActor
LogRequestActor had been originally constructed to send all updates
directly to the persistence layer.  In catchup mode, however, that won't
work -- those updates need to come in through the PaxosStateBridge so
that they can be accounted for in the eventBuffer.

This parameterizes all log request messages to include a target
ActorRef: where to send the final OrderedEvents once they've been
received. Also adds the target = PaxosStateBridge for requests sent
during catchup mode.

Depends on "request gaps" commit for the addition of targetActor in
PaxosStateBridge.

Change-Id: I1de4a8c701b1c11737c950b4ce137af6cdc71c6e