Fixed a bug related to piping multiple lines of log content into walt for parsing.
Fixed a bug where walt did not commit to the db
being created a remainder of rows.

Changeset: Running this on dstr04 revelaled that
earlier versions of Python (less than 2.7)
don't have OrderedDict. I added that
with this commit.

walt.py can now run in dev in my home dir and
is ready for packaging as an executable
with freeze.

Change-Id: I3884f116a261f029abbd8790a2aa8d34592bb8b2