Apply writes synchronously when bootstrapping
The async approach for applying writes when bootstrapping is causing
problems with application startup.  The bootstrapping appears to be
completed, and the StateAgent is informed that the PersistenceState is
Initialized, but the changes are still queued up to be applied to the
RequestHandler.  This leads to the application being started without the
brain being initialized.

This change applies the changes synchronously.  It isn't terribly
efficient, causing bootstrapping to take about twice as long as it
appeared to before, but it does guarantee that the brain is fully
initialized when the SiriusSupervisor reports that initialization is
complete.

Change-Id: I47a53af56549776c0f6f6046ef4b572bfe89b389