Synchronous log replay application
StateSup replays log during initialization, such that the state and
persistence layers get fully ready copies of objects.

With this we only need to wait for StateSup to be initialized,
consider this a TODO for when we want to tackle that.

Also, I know the companion constructor is getting a little busy.

Change-Id: I0c3d48a490538dae35adfa28bfe61a05fb9ea294
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/3885
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>
Tested-by: GerritAHP