Make Base64Codec thread safe
The alternative is to pass a fresh SerDe into the log requesting
business things each time we spin them up, but we don't actually
know the given impl due to the current construction.  This should
be a good fix for now until we replace the log with the more
efficient format.

Change-Id: Ie1a85e7ec5a3f18e357ed29c304827e58ab8f2a0
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/4197
Tested-by: GerritAHP
Reviewed-by: michael bevilacqua-linn <michael_bevilacqua-linn@comcast.com>