Index file repair
This is covered by the integration test for the missing index file
case, but not the cut off one, it should work fine there though.  We
will see if we can figure out a nice way to test this in the morning,
but if you want to +2 it no big deal :).

Change-Id: Ibe2a910fa6164d26bbfa8f6a10ff5c81367fe38c
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/4281
Reviewed-by: michael bevilacqua-linn <michael_bevilacqua-linn@comcast.com>
Tested-by: michael bevilacqua-linn <michael_bevilacqua-linn@comcast.com>