Add ability to directly query state
This makes the test harness stuff much easier- presently to retrieve
a subrange from a node you would need to send a RequestFromRemote
message to a node, with its address as the "remote", and your address
as the replyTo.  A session is then established between the node and
itself, and then messages are piped over the wire.  It's much easier
to just say "give me what you have" directly to a node.  Note, we
don't want to request too massive ranges, capping it at 1k for now.

Change-Id: I3d5bcc4d3d19852ae791c6ad003c02a4758c051a
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/4455
Tested-by: GerritAHP
Reviewed-by: trevor menagh <trevor_menagh@comcast.com>