Track proposals, repropose if necessary
Track sent proposals in the Replica by their slot number.  If a decision
arrives for a particular slot number and our tracked proposals contains
that slot number as a key, check whether the incoming decision's command
is the one we proposed -- in which case it can be safely dropped -- or
if it's a different command, in which case we'll need to repropose our
stored proposal with a new slot number.

Reaps proposals from local proposal list that are out of reapWindow.

Change-Id: I118253efe4ce36675622c8cfd906688185a4d203
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/4696
Reviewed-by: michajlo matijkiw <michajlo_matijkiw@comcast.com>
Tested-by: GerritAHP