Remove PaxosSup.Submit message
The original idea behind this was to encapsulate requests in a type
so that we could easily add stuff, namely a retry count. However, it
wound up just creating some extra cruft in SiriusSupervisor, and we
decided to go the way of time-bounded retries, so we'll drop it for
now.

Change-Id: I6f1f03258295e8fbb91808a5edd3c867570f037e
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/4699
Tested-by: GerritAHP
Reviewed-by: michael bevilacqua-linn <michael_bevilacqua-linn@comcast.com>