Create a trait for registering monitors from within an actor
Mix this trait in to register an MBean local to an actor, takes
care of any possible name space collisions, assuming the caller is
using such even mildly intelligently (if you register the same mbean
twice then that's stupid).

MBeanServer is passed down through SiriusConfiguration, the idea is
that if you attach an MBeanServer to SiriusConfiguration, this trait
will register mbeans when called, if not it will do nothing.  Since
the mbean param is call-by-name, it actually won't even be created,
which is nice.

Change-Id: I13fb039c2f2c8a0c2f49083d7e58f9cc2924a088
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/4752
Tested-by: GerritAHP
Reviewed-by: karl martino <karl_martino@comcast.com>