Wire up some Leader monitors
Propogate the MBean stuff down

It's a little weird how we have to do this because of the
restrictions on MBeans and the need to close over the scope of the
class.  Some alternatives would be:

  - make Leader implement LeaderMBean (don't really feel to hot about
        this)
  - put the MBean stuff in another package and have the constructor
        take call by name params so we can give them the scope of
        the local vars

All alternatives are strange, but none of them are particularly
difficult to do, so it's my opinion that we just keep going as is for
not and see if we decide that we need to change.

Change-Id: I758f356e3abae51fef4e7db20a9cb230a36ca56a
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/4753
Tested-by: GerritAHP
Reviewed-by: trevor menagh <trevor_menagh@comcast.com>