Catchup actor for client-state-only gap fetching
This actor can both fill gaps and proactively seek updates that its
parent doesn't know about.  It starts with a "firstGapSeq", basically
highestSeqSeen + 1, and starts requesting chunks from there.  As long as
it's getting full chunks, it keeps requesting.  If it gets a non-full
chunk (ie, the target only has part of the request, as the tail is into
the future...) then it stops after this chunk.  If its request is timed
out, it stops.

This actor is designed to be called on a schedule from the
PaxosStateBridge.  The bridge should check on a schedule to see whether
the GapFetcher is still running, and if so, assume it's doing its job.
Otherwise, fire up a new one.  This functionality incoming in future
commit.

Change-Id: Id654b44b14a8877ba131cae9f73a64ead0adeadc
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/4863
Reviewed-by: trevor menagh <trevor_menagh@comcast.com>
Tested-by: GerritAHP