Wire in state-free catchup
State-free catchup involves the PaxosStateBridge spinning up a
GapFetcher actor, which proactively asks remote nodes for log chunks.
Based on the local node's current sequence number, it asks for batches
(right now set at 1000) of events, and as long as it's getting back a
full batch, asks for the next.  When it receives nothing (timeout) or a
partial batch (assumes the remote node doesn't have anything beyond
that) the GapFetcher terminates quietly.

This removes the need for LogRequest and LogSending/Receiving actors.

Also: Updated membershipHelper to hang onto stateAgent and localRef, so
that it could be more cleanly injected into classes that would otherwise
have to keep references to both of those.  Wired this into the
PaxosStateBridge.  This will probably be refactored a little more, but
will wait until LogRequest*Actor are gone.

Next: ITEST!
Then: unwire LogRequest*, delete those files.

Change-Id: I476da3ae88ab694bef7ddc6bce81d1f153bada24
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/4936
Reviewed-by: trevor menagh <trevor_menagh@comcast.com>
Tested-by: trevor menagh <trevor_menagh@comcast.com>