Rearrange uberstore package a bit
Peter and I have been discussing this for a while, but the code at
the time was moving too fast for this to be a sane feat. Now that
things have slowed down in this package it's cool to go forward with.

No code changes, just rearranging, the outside world should have no
idea that this happened.

  - Move all data stuff under the data package
  - Moved all checksum stuff to common
  - updated imports to reflect changes

Change-Id: I0fc9890743253979066a27e03cfce7ca42a0925b
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/4933
Tested-by: GerritAHP
Reviewed-by: walter eggert <walter_eggert@comcast.com>