Add connection timeout and backoff to akka config
It appears that we may get ourselves into situations sometimes where
we try way too hard to send a message to a node that is clearly down.
The result is that the sending node slows down, because it keeps
trying to send, and the receiving node gets a flood of messages when
it comes back online because the sending node finally figured itself
out.

At least I think this is the case, this config should handle that.

Change-Id: I5d71cd606328e7abe4dd88659886589123ade2a6
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5051
Tested-by: GerritAHP
Reviewed-by: walter eggert <walter_eggert@comcast.com>