Up slf4j to 1.7.0
Starting with 1.6.0 the no-op MDCAdapter is used if no other binders
are found. 1.7.0 because it's badass.  Before sirius would crash
hard, taking down the JVM with it per akka's configuration, and with
good cause- it couldn't load a class it expected to be able to load!

I figure that if we are shipping sirius with slf4j as the default
logger we should make sure that non-configuration will not kill the
end user.

Change-Id: I4d7fc68d386adf4ea92378522a2c5ec8912cfadf
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5053
Tested-by: GerritAHP
Reviewed-by: walter eggert <walter_eggert@comcast.com>