Allow external akka configuration to be passed in
Allow clients to tweak config without needing a full on new build of
Sirius.

Also clean up the interactive stuff that depends on it.

Now host/port config overrides all, followed by any external config,
followed by the config supplied in the jar as it was before. The
file system is first checked for external config, then the classpath.

Change-Id: I9220720b12f37317e835e6560383ca9d54504e0b
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5055
Tested-by: GerritAHP
Reviewed-by: walter eggert <walter_eggert@comcast.com>