Commander timeout nullifies commandered slot
We observed that while the replica was doing the right thing
reusing a slot number for which it got no feedback.  However, it
turned out those reproposed slots weren't making any progress.
Looking into it deeper it appears the likely cause is that a
commander for that slot would time out, then subsequent proposals
to the leader for that slot would fail because the leader already
had a proposal for that slot.  Here we add some feedback from the
commander when it times out to let the leader no to release its
proposal for that slot.  This should be reasonably rare.

Change-Id: I8648d86604387f21f86927a0b38b2e057efa5ea6
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5110
Tested-by: GerritAHP
Reviewed-by: walter eggert <walter_eggert@comcast.com>