Eliminate createIterator in favor of foldLeftRange
Publicizes foldLeftRange.  In retrospect, these could have been done
separately.  This actually rips out createIterator entirely, which I
think we're fine with... folding is the pattern that works best for most
of our applications.  We may want to consider adding a foreach to fill
in the rest.

After this, we can delete LogIteratorSource, since it's no longer used
at all by the rest of the application.

patchset2: Addressed some comments from patchset1
patchset3: addressed the fact that I forgot to run tests for patchset2

Change-Id: Iec25df5f609fbd01376f8b5c0e6d203e71c2a37f
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5116
Tested-by: GerritAHP
Reviewed-by: michajlo matijkiw <michajlo_matijkiw@comcast.com>