Change MBean name key type -> name
So here's the deal... SevOne does this thing where it looks for a
"name" field in an MBean's name, and will use that instead of the
full object name if it's available.

Here's where it gets weird, we have a "sysname" field in the object
name, and SevOne decided to use this field to identify the MBean in
its UI.  This resulted in having a whole bunch of things labeled as
"sirius-system,[random number]" in the SevOne API.

So here's my theory, SevOne decides object name the following way:

  s/name=\([^;]*\)/\1/

So, it sees sysname, and decides that it will use that.

Throw a name field into those kvs, and hope that it gets parsed
instead of sysname.

What a piece...

Change-Id: I90e945e5771cbc2e69f9c2ca0e0a558bbb5def1f
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5203
Tested-by: GerritAHP
Reviewed-by: trevor menagh <trevor_menagh@comcast.com>