Fix replica DecisionHint reaping bug
Also make ReapWindow configurable in the replica.  Did that as a part of
debugging what ended up being the DecisionHint bug, and it didn't seem
worth it to disentangle the two into separate commits.

This should fix some of the badness we've been experiencing in the test
harness, and fixes a bug that was causing proposals (that may or may not
have ever actually been commandered) to vanish, seemingly randomly.

Change-Id: I80a8ccbd21c598f582c35029e32079bc779df777
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5207
Reviewed-by: michajlo matijkiw <michajlo_matijkiw@comcast.com>
Tested-by: GerritAHP