Revamped FullSystemITest
Reworked the itest that exercises the whole system in a few ways.  In
particular, added a mechanism to fire random events buckshot-style
across all available nodes as well as using UberStore as the backend
SiriusLog format.

Added 2 new tests to really work the system.  One for a happy case with
50  messages pouring in, another showing catchup.  There should be
plenty more we can think of for this.

patchset2: toned down the itest a little.

Change-Id: Ie38196c52c0442082a947eddb9588690ae59e08c
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5211
Tested-by: GerritAHP
Reviewed-by: michajlo matijkiw <michajlo_matijkiw@comcast.com>