Slim down SiriusInfo
Removed membership and next sequence from SiriusInfo, they're exposed
elsewhere.

Expose node address and isTerminated (not online, but terminated) in
SiriusInfo.

The idea here is that SiriusInfo will stick around even if the node
crashes- when a node shuts down it takes its monitors with it. This
guy sticks around until the node is terminated as a reminder that
the node was once there.

Change-Id: Ie5f3342f300b1fa1674654a902de1b361430efe6
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5251
Tested-by: GerritAHP
Reviewed-by: peter cline <peter_cline2@comcast.com>