Create PaxosStateBridge in SiriusSupervisor
It was created within the context of the supervisor before, but we
did some trickery with implicit scopes to hide it so that we could
do the naive ordering actor vs paxos bit, and just because that's
how the code wound up evolving.

It's time for the state bridge to come out of it's shell and shine
as the top level subordinate it is.

Change-Id: Ia28a568d780fb9e1fd284fb7217dc568b925671b
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5262
Tested-by: GerritAHP
Reviewed-by: peter cline <peter_cline2@comcast.com>