Bandaid for FullSystemITest
We believe we have indentified proposals going MIA being a symptom
of catch up happening and outstanding proposals in need of reproposal
getting dropped on decision hint.

This will delay catch up long enough for the test to finish (and in
reality we probably want the delay anyway).  We do however inject
a catch up into the state bridge for the catch up test so that we
don't need to wait the full 30 seconds.

The test solution probably isn't the prettiest (a better solution is
proposed in a comment), but I think this wins the time/effort trade
off at the moment, especially because this is blocking builds.  If
I get the time tonight I'll try to implement the better solution.

Change-Id: Ie796ed876e7e296919c10790e0f6d1b216b90886
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5273
Tested-by: GerritAHP
Reviewed-by: walter eggert <walter_eggert@comcast.com>
Reviewed-by: peter cline <peter_cline2@comcast.com>