Introducing RichJTreeMap
Yesterday we made the observation that we have a lot of similar code
in the PSB, Leader, Replica, and Acceptor- namely they were all using
a JTreeMap in similar ways.  This aims to allow us to exploit that
commonality and reduce code, with the end goal that less code means
it's easier to find and fix bugs.

Change-Id: I00e0356bc38b00996db128396d3e41cbc95b15c4
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5287
Tested-by: GerritAHP
Reviewed-by: peter cline <peter_cline2@comcast.com>