Add master/slave mode based on cluster membership
A supervisor periodically checks the membership and wonders, "am I still
in the cluster?"  If the answer is yes, the supervisor verifies that the
ordering actor is alive and happy and life continues.  If the answer is
no, it goes into sort of a "slave" mode, turning off its ordering actor
(which is the paxos subsystem), while leaving the PaxosStateBridge
running, which will do constant catchup as normal.

This also makes the state bridge fetch frequency configurable.

OrderingActor now starts off as None, and stays so until it is updated
on schedule.  The MembershipActor updates the MembershipAgent on
startup, and this is fed into the supervisor, which immediately checks
for paxos participation.

Fixed a number of things, including root causes for test failures.  Also
removed commander retry, which will go in as a separate commit.

(aka Peter Cline's war against determinism)

Change-Id: I7ddf9e4f4b6a20a4bf7f636e036e9f5eb9f2bcc8
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5315
Tested-by: GerritAHP
Reviewed-by: michajlo matijkiw <michajlo_matijkiw@comcast.com>