Add Commander retries on Timeout
This is done by starting a commander with a number of retry "ticks,"
which it returns to the leader with its timeout message.  When the
Leader receives the timeout message, if there are ticks remaining, it
restarts the commander.  Otherwise, it lets it die and removes the
proposal from its active proposals.

Addressed mishu's comments from previous patchset.

Change-Id: I1dd3547af7e62d34354de9174f6de6148420710d
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5396
Reviewed-by: michajlo matijkiw <michajlo_matijkiw@comcast.com>
Tested-by: GerritAHP