Suppress RemoteClient* and RemoteServer* logging
This stuff was spewing tons of nonsense, making debug difficult and
overwhelming the logs. If we really want to get info on it we can
create a handler to output it (note there may be a lot, one for
every failed send), or demote it to a lower log level.

Change-Id: I8137a21f8014c38c59a8e247f34502646f218aef
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/5451
Tested-by: GerritAHP
Reviewed-by: peter cline <peter_cline2@comcast.com>