Use a conforming artifact name for SBT.
The SBT build was creating an artifact in a target/scala-2.9.2 directory.
The artifact was also named sirius_2.9.2-1.0.jar, which did not conform to
the naming conventions used in the POM or in other builds.

These changes tell SBT to stop creating the scala-2.9.2 directory, and also
name the .jar file as target/sirius.jar, which matches what is created via mvn.

Change-Id: I7e08ea4d1a3e125ffe0d4cd247be8e6acfc1ea1d
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/6172
Tested-by: GerritAHP
Reviewed-by: trevor menagh <trevor_menagh@comcast.com>