Refactoring of failure interpretation in FullSystemItest
Trying to figure out why we get spurious failures, I though this
chunk of code was a little misleading before, proposing that this
may be a little cleaner.

By the way, the test failures appear to be from a leader being cut
down as part of a test, and then requests being shot to it, and
other nodes forwarding to nowhere. Reproposed slots aren't getting
reaped quickly enough, so they just go mia at the end of the test.

Change-Id: I53b9b721c60da418e0e4a6ba5fda8ef70696269a
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/6183
Tested-by: GerritAHP
Reviewed-by: walter eggert <walter_eggert@comcast.com>
Reviewed-by: peter cline <peter_cline2@comcast.com>