Eliminate spurious test failures forever
Ha ha, just kidding. However, added a bit to FullSystemITest that
triggers leader election and waits a couple of seconds for it to occur
before going into master/slave mode.  The master/slave test has been
failing, I (we) think, because we removed the leader from the membership
(put the leader into slave mode) and then fired a bunch of events at
other nodes, which forwarded it to the slave (which subsequently dropped
them on the floor).  This change should hopefully avoid that situation.

Change-Id: I79d5814be94b01593d1d0f556fa76a6fb8eb6807
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/6356
Reviewed-by: walter eggert <walter_eggert@comcast.com>
Tested-by: GerritAHP
Reviewed-by: trevor menagh <trevor_menagh@comcast.com>