Fix bug where two-pass compact didn't like empty log
UberTool#twoPassCompact would fail if no events were gathered for
keeping in the first phase due to calling next on an empty iterator.
Make that not happen anymore.  Factored the phases into two local
private methods for better readability.

Seems a little busy with the extra private helpers, but I don't think
it's so bad, plus this code is likely to be replaced soon by more
clever live compaction logic.

Change-Id: Ib140a92afc0258ad7a95a427ac06cc2b326c6a12
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/12469
Tested-by: GerritAHP
Reviewed-by: walter eggert <walter_eggert@comcast.com>