Make UberStore constructor private
Now always instantiate via companion object. Doing this to reduce the
number of ways we can create an UberStore to one in preperation for
more complex live compaction structuring. Sacrificed ReadOnlySeqIndex
to improve simplicity, so `wailtool tail -f` will no longer work,
I think it's probably worth forgoing this functionality for now in
favor of simplicity of the interface.

Change-Id: I325b4501641ea1a4deb7d03e2a54f1bff5a50be3
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/13611
Tested-by: GerritAHP
Reviewed-by: ilirjan papa <ilirjan_papa@comcast.com>
Reviewed-by: peter cline <peter_cline2@comcast.com>