Disk only based SeqIndex, no cache
Naming isn't the best WRT to the other implementation, but you get
the idea. An index file implementation which is super light on memory
and hopefully quite performant as well.

Tests are integration style tests, because testing this with mocks
would be prohibitively difficult, and as so would yield crappy
results.

Not wired up anywhere yet, just an orphan waiting for its time to
shine.

Change-Id: I1b689c7b0e029b87aca4f690bed436c4dcf878c7
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/13989
Tested-by: GerritAHP
Reviewed-by: michael bevilacqua-linn <michael_bevilacqua-linn@comcast.com>