UberPair -> UberDir, new dir based log approach
Breaking changes to current uberstore organization with the vision
for a brighter tomorrow.  Dir based organization as opposed to file,
it is with great shame that I admit that we should have had the
foresight for this at uberstore's inception.

The plan is to introduce an "UberShim" at some point between now
and the next release which will transition the old format to the
new.

Change-Id: I34533f28d2729cf0c778f79195552336c2ea36b8
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/207
Tested-by: trevor menagh <trevor_menagh@comcast.com>
Reviewed-by: trevor menagh <trevor_menagh@comcast.com>