Measure and monitor round-trip time between members
This allows us to periodically fire off ping messages to all known
members, recording the round-trip time when the Pong is received. It
also records the "last Pong received" for each member, which could be
useful for detecting mailbox issues / dead members.

Right now, it's just dumping out the maps, which is not in a very useful
format for anything other than people looking at it.

Change-Id: Ib97efbb93ca06937b2a7e7d10ee5dcc89a1c3fc4
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/1975
Tested-by: GerritAHP
Reviewed-by: trevor menagh <trevor_menagh@comcast.com>