Add monitoring around LeaderPing process
JMX monitors around the leaderPings.  We spin up an actor to ping the
last-known elected leader, and to keep doing so as long as the leader is
alive and replying. We've seen situations, though, where a node has gone
bad, but still seems to be responding to ping. At least, we think so: we
can only infer. This gives us some insight into those actual responses
so we can verify that the functionality is actually working as expected.

Change-Id: I9259990a31d892e3551227000ae36bcd6f741117
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/2865
Tested-by: GerritAHP
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>